# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cigar Helper API",
    "description": "API для управления хьюмидорами и сигарами",
    "contact": {
      "name": "Cigar Helper Team"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Получает список всех обзоров",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "Фильтр по пользователю (опционально)",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cigarBaseId",
            "in": "query",
            "description": "Фильтр по каталожной сигаре (CigarBase)",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userCigarId",
            "in": "query",
            "description": "Фильтр по записи коллекции (UserCigar)",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewListItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Создает новый обзор",
        "requestBody": {
          "description": "Данные для создания обзора",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Reviews/{id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "Получает детальную информацию об обзоре по его ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID обзора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Reviews"
        ],
        "summary": "Обновляет существующий обзор",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID обзора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Данные для обновления обзора",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "summary": "Мягкое удаление обзора автором (строка остаётся в БД, публично не отображается; восстановление — staff через `/api/admin/reviews/.../restore`).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID обзора",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateReviewImageRequest": {
        "required": [
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "imageUrl": {
            "minLength": 1,
            "type": "string"
          },
          "caption": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateReviewRequest": {
        "required": [
          "cigarBaseId",
          "content",
          "rating",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 3,
            "type": "string"
          },
          "content": {
            "minLength": 1,
            "type": "string"
          },
          "rating": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "cigarBaseId": {
            "type": "integer",
            "format": "int32"
          },
          "userCigarId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "smokingExperience": {
            "type": "string",
            "nullable": true
          },
          "aroma": {
            "type": "string",
            "nullable": true
          },
          "taste": {
            "type": "string",
            "nullable": true
          },
          "construction": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "burnQuality": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "draw": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "smokingDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "smokingDurationMinutes": {
            "maximum": 720,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bodyStrengthScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "aromaScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pairingsScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateReviewImageRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "isAuthorProfilePublic": {
            "type": "boolean"
          },
          "userAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "cigarBaseId": {
            "type": "integer",
            "format": "int32"
          },
          "userCigarId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cigarName": {
            "type": "string",
            "nullable": true
          },
          "cigarBrand": {
            "type": "string",
            "nullable": true
          },
          "cigarCountry": {
            "type": "string",
            "nullable": true
          },
          "cigarLengthMm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cigarDiameter": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cigarWrapper": {
            "type": "string",
            "nullable": true
          },
          "cigarBinder": {
            "type": "string",
            "nullable": true
          },
          "cigarFiller": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewImageDto"
            },
            "nullable": true
          },
          "smokingExperience": {
            "type": "string",
            "nullable": true
          },
          "aroma": {
            "type": "string",
            "nullable": true
          },
          "taste": {
            "type": "string",
            "nullable": true
          },
          "construction": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "burnQuality": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "draw": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "bodyStrengthScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "aromaScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pairingsScore": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "smokingDate": {
            "type": "string",
            "format": "date-time"
          },
          "smokingDurationMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "imageBytes": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "caption": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "isAuthorProfilePublic": {
            "type": "boolean"
          },
          "userAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "cigarName": {
            "type": "string",
            "nullable": true
          },
          "cigarBrand": {
            "type": "string",
            "nullable": true
          },
          "cigarBaseId": {
            "type": "integer",
            "format": "int32"
          },
          "mainImageBytes": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "imageCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateReviewRequest": {
        "required": [
          "content",
          "rating",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 3,
            "type": "string"
          },
          "content": {
            "minLength": 1,
            "type": "string"
          },
          "rating": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "smokingExperience": {
            "type": "string",
            "nullable": true
          },
          "aroma": {
            "type": "string",
            "nullable": true
          },
          "taste": {
            "type": "string",
            "nullable": true
          },
          "construction": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "burnQuality": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "draw": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "nullable": true
          },
          "smokingDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "smokingDurationMinutes": {
            "maximum": 720,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bodyStrengthScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "aromaScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pairingsScore": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "imagesToAdd": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateReviewImageRequest"
            },
            "nullable": true
          },
          "imageIdsToRemove": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}